requirements.txt
azure-cognitiveservices-speech==1.34.1
azure-storage-blob==12.19.0
bs4==0.0.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Faker==22.2.0
google-cloud-firestore==2.14.0
google-cloud-aiplatform==1.43.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.15.1
google-cloud-translate==3.14.0
jieba==0.42.1
zhon==2.0.2
matplotlib==3.8.2
pandas==2.1.4
pillow==10.2.0
plotly==5.18.0
pydub==0.25.1
pymongo==4.6.1
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
spacy==3.7.2
streamlit==1.33.0
streamlit-elements==0.1.0
streamlit-mic-recorder==0.0.4
st-annotated-text==4.0.1
moviepy==1.0.3
werkzeug==3.0.1
langdetect==1.0.9
firebase-admin==6.4.0
scipy==1.12.0
langsmith==0.1.23
langchain-community
langchain-experimental
langchain-google-vertexai==0.1.0
langchain==0.1.11
numexpr==2.9.0
google-search-results==2.4.2
numexpr==2.9.0
sympy==1.12
langgraph==0.0.26
pytesseract==0.3.10
opencv-python==4.9.0.80
scikit-image==0.22.0